boston-house-price-prediction/requirements.txt
torch
torchvision
pandas
progressbar2
sklearn